+++ test/unittests.cpp
@@ -33,7 +33,7 @@ protected:
     int vocab_size = 500;
     Tokenizer EmptyTokenizer = Tokenizer(name, vocab_size);
     Tokenizer TestTokenizer = Tokenizer(name+"-filled", vocab_size);
-    std::list<unsigned long long int> test_encoded_text = {586, 606, 607, 617, 1, // This
+    std::list<uint64_t> test_encoded_text = {586, 606, 607, 617, 1, // This
                                                            607, 617, 1,  // is
                                                            599, 1,  // a
                                                            618, 603, 617, 618, 1}; // test
@@ -57,7 +57,7 @@ TEST_F(TokenizerTest, test_build_vocab) {
 
 
 TEST_F(TokenizerTest, encoder_test) {
-    std::list<unsigned long long int> encoded_text = TestTokenizer.encode("This is a test");
+    std::list<uint64_t> encoded_text = TestTokenizer.encode("This is a test");
     EXPECT_EQ(test_encoded_text, encoded_text);
 }
